Refactor HTML parser to handle multiple websites
This commit renames the HTML_Parser.cs file to Data_Scraper.cs and refactors the code to allow for parsing of multiple websites. The htmlParser function now takes a URL parameter, which is used to load the website's HTML content using HtmlAgilityPack. A new getWebsiteName function has been added to extract the website name from the URL. The parsed HTML content is saved in a file named after the website name in a "scraped_html" folder.